feat: security updates
Source: 04f2847bf33a0bd1cd6af14c7b24da95ad824aed

+++ pages/api/teams/[teamId]/invitations/index.ts
@@ -81,6 +81,10 @@ export default async function handle(
         return res.status(403).json("You are not part of this team");
       }
 
+      if (userTeam.role !== "ADMIN") {
+        return res.status(403).json("Only admins can revoke invitations");
+      }
+
       // delete invitation
       await prisma.invitation.delete({
         where: {
